# swagger.yaml
openapi: 3.0.3
info:
  title: MediGo API
  description: API for managing the 11-minute medicine delivery platform, including user registration, order processing, and delivery management.
  version: 1.0.0
servers:
  - url: http://localhost:8081/v1
    description: Local server for testing
components:
  schemas:
...
    UserBase:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - user_id
        - name
        - email
        - created_at
        - updated_at

    # UserCreate Schema (Used for creating a new user)
    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
      properties:
        password:
          type: string
      required:
        - password

    # UserUpdate Schema (Used for updating an existing user)
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        updated_at:
          type: string
          format: date-time

    # User Schema (Used for returning a user object)
    User:
      allOf:
        - $ref: '#/components/schemas/UserBase'
      properties:
        user_id:
          type: string
          format: uuid
      required:
        - user_id
    # EntityBase Schema (Base schema for entity data)
...
paths:
...
  /users/:
    post:
      summary: Create a new user
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    get:
      summary: Get a user by user ID
      operationId: get_user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      summary: Update an existing user
      operationId: update_user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    delete:
      summary: Delete a user by user ID
      operationId: delete_user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted successfully"
        '404':
          description: User not found
...